feat: UI优化 - 产品更名SnapVid + 背景动画 + 平台真实Logo + footer署名qikunma
- 产品名从ytdlp-web改为SnapVid，slogan: 一键抓取，极速下载
- 添加渐变背景浮动动画效果
- 平台列表图标替换为真实SVG品牌Logo
- Footer署名改为Made by qikunma
- 修复Bilibili URL自动补全(缺少https://www.前缀)

+++ backend/api/schemas.py
@@ -31,7 +31,7 @@ class VideoInfoResponse(BaseModel):
     """Response model for video info."""
 
     title: str = Field(default="", description="Video title")
-    duration: Optional[int] = Field(default=None, description="Duration in seconds")
+    duration: Optional[float] = Field(default=None, description="Duration in seconds")
     duration_string: str = Field(default="", description="Human-readable duration")
     thumbnail: str = Field(default="", description="Thumbnail URL")
     uploader: str = Field(default="", description="Uploader name")
